Add GeoResourceController concern for DRYer code
