+++ spec/controllers/scanned_maps_controller_spec.rb
@@ -301,6 +301,8 @@ RSpec.describe ScannedMapsController do
     end
   end
 
+  # Tests functionality defined in `app/controllers/concerns/geo_resource_controller.rb`
+  #   Acts as a 'master spec' in this regard
   describe "PUT /concern/scanned_maps/:id/extract_metadata/:file_set_id" do
     let(:user) { FactoryGirl.create(:admin) }
     let(:file) { fixture_file_upload('files/geo_metadata/fgdc.xml', 'application/xml') }
